🚚 Rename add-new to add-menu
