+++ src/lib/routes.js
@@ -1,6 +1,6 @@
 import { renderIndex } from '../pages'
 import { renderResult } from '../pages/result'
-import { renderAddMenu } from '../pages/add-new'
+import { renderAddMenu } from '../pages/add-menu'
 
 const routes = {
   '/': renderIndex,
